+++ .github/a11y-baseline.json
@@ -1054,6 +1054,11 @@
       "helpUrl": "https://dequeuniversity.com/rules/axe/4.12/color-contrast?application=playwright"
     },
     {
+      "key": "Tokenizer::Touch: disabled, with a reason::color-contrast",
+      "impact": "serious",
+      "helpUrl": "https://dequeuniversity.com/rules/axe/4.12/color-contrast?application=playwright"
+    },
+    {
       "key": "Toolbar::Composition: Tab Navigation::landmark-unique",
       "impact": "moderate",
       "helpUrl": "https://dequeuniversity.com/rules/axe/4.12/landmark-unique?application=playwright"
